update db and tasks
